requirement.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
Markdown==3.8.2
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
